requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
djangorestframework==3.12.4
mysql-client==0.0.1
mysql-connector-python==8.0.26
mysqlclient==2.0.3
protobuf==3.17.3
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
